Debugging of sql service. Target platform update icu

+++ at.sunplugged.z600.core.conveyor/src/at/sunplugged/z600/conveyor/api/ConveyorPositionCorrectionService.java
@@ -6,9 +6,9 @@ public interface ConveyorPositionCorrectionService {
 
     public void stop();
 
-    public double getRuntimeLeft();
+    public long getRuntimeLeft();
 
-    public double getRuntimeRight();
+    public long getRuntimeRight();
 
     public void setRuntimeLeft(long ms);
